refactor(lotto): 여러 도메인 객체를 사용해야 하는 작업은 특정 도메인 내부가 아니라 서비스 레이어에서 작업하도록 수정
