+++ src/main/java/lotto/domain/ScoreInfo.java
@@ -1,8 +1,6 @@
 package lotto.domain;
 
-import lotto.vo.Lotto;
 import lotto.vo.Score;
-import lotto.vo.WinningInfo;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -23,36 +21,7 @@ public class ScoreInfo {
         return scoreInfo.get(score);
     }
 
-    public void calculateRank(Lotto lotto, WinningInfo winningInfo) {
-        int matchCount = getMatchCount(lotto, winningInfo);
-
-        if (matchCount == 6) {
-            addScore(Score.FIRST);
-        } else if (matchCount == 5) {
-            if (isBonusMatching(lotto, winningInfo)) {
-                addScore(Score.SECOND);
-                return;
-            }
-            addScore(Score.THIRD);
-        } else if (matchCount == 4) {
-            addScore(Score.FORTH);
-        } else if (matchCount == 3) {
-            addScore(Score.FIFTH);
-        }
-    }
-
-    private boolean isBonusMatching(Lotto lotto, WinningInfo winningInfo) {
-        return lotto.contains(winningInfo.getBonus());
-    }
-
-    private static int getMatchCount(Lotto lotto, WinningInfo winningInfo) {
-        return (int) lotto.getNumbers()
-                .stream()
-                .filter(number -> winningInfo.contains(number))
-                .count();
-    }
-
-    private void addScore(Score score) {
+    public void addScore(Score score) {
         scoreInfo.put(score, scoreInfo.get(score) + 1);
     }
 }
